requirements.txt
azure-appconfiguration==1.5.0
azure-appconfiguration-provider==1.1.0
azure-core==1.30.0
azure-cosmos==4.3.1
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
Flask==2.0.3
Werkzeug==2.2.2
flask-cors==4.0.0
